Server/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
psycopg2-binary==2.9.5
requests==2.28.2
PyJWT==2.6.0